For #32: Clean up the search logic for hardware directories
Moves the search logic for boards.txt and programmers.txt files into a
central location. Also makes it more clear how we're parsing the package
and arch out of the filepath. Additionally, added a slight hack to also
search /etc/arduino with the default package and arch of 'arduino' and
'avr'. I'm not sure this is correct, but it seems to be what is needed
from the attached issue.